Update composition types on AcceptDisplayChanges
When AcceptDisplayChanges is called, the layer composition type
on the client side is deemed to have been changed to the
composition changes that were requested. Update the device state
accordingly.

Move the configuration of the layer flags to BuildLayerStack so
that the client accepted composition type is accounted for.

Bug: 30880078
Bug: 30119044
CRs-Fixed: 1059410
Change-Id: I09eab9dcc8d169acc761e84a6c76712712c2b934